+++ application/rest-api-adapter/src/main/java/onlydust/com/marketplace/api/rest/api/adapter/ApplicationsRestApi.java
@@ -6,7 +6,7 @@ import lombok.AllArgsConstructor;
 import onlydust.com.marketplace.api.contract.ApplicationsApi;
 import onlydust.com.marketplace.api.rest.api.adapter.authentication.AuthenticatedAppUserService;
 import onlydust.com.marketplace.project.domain.model.Application;
-import onlydust.com.marketplace.project.domain.port.input.UserFacadePort;
+import onlydust.com.marketplace.project.domain.port.input.ApplicationFacadePort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -19,12 +19,12 @@ import static org.springframework.http.ResponseEntity.noContent;
 @AllArgsConstructor
 public class ApplicationsRestApi implements ApplicationsApi {
     private final AuthenticatedAppUserService authenticatedAppUserService;
-    private final UserFacadePort userFacadePort;
+    private final ApplicationFacadePort applicationFacadePort;
 
     @Override
     public ResponseEntity<Void> deleteApplication(UUID applicationId) {
         final var authenticatedUser = authenticatedAppUserService.getAuthenticatedUser();
-        userFacadePort.deleteApplication(Application.Id.of(applicationId), authenticatedUser.getId(), authenticatedUser.getGithubUserId());
+        applicationFacadePort.deleteApplication(Application.Id.of(applicationId), authenticatedUser.getId(), authenticatedUser.getGithubUserId());
         return noContent().build();
     }
